pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "weather-cmd"
version = "0.1.0"
description = "Terminal weather app with forecast graphs"
requires-python = ">=3.12"
dependencies = [
    "textual>=1.0.0",
    "textual-plotext>=1.0.0",
    "textual-image>=0.6.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
weather-cmd = "weather_cmd.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/weather_cmd"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
